input handler - fixed le bug
I was using setTimeout to come back and delete the displayed html element..  but I was jsut calling the normal display method..  which meant if I double clicked, I'd get a hit for "don't delete yet.." which caused ANOTHER setTimeout call!  The solution was to create a new method just for that call.  clearlyOverlayElement