HillBuilder bug, SceneLayer vertical location/height in game window
In hillBuilder, hills start at 1..  but my little repeat check was like "if we are at last number, wrap back to zero.".  WOOPS, needed to wrap back to 1.

SceneLayer - easy one, dunno why i was trying to overthink it.  it now sets the scene height (and bottom edge) based on the passed in relative_height.