more scene layers!  horizon support added to layers! Z-INDEX ISSUE!
Gadddmammmitt!  So I had my layer z-indexes set to -15, -10, -5, 0.  (Front pasture is zero.).  My canvas layer is z-index 10.  Well!  My pasture was drawing over the canvas layer...  I set pasture to -1, and it was drawing behind the canvas.  SO!  Does that mean z-index only works with negative numbers??  It seems to be...  ALSO, z-index is whole numbers only.  Just saying!