SceneScaler - OK, I think location works!
Now I just need to do the scaling.

NOTE: with the z-Index values I'm using at the moment, I need to round/floor that value..  If there are 2 things near each other, and one is at like 1.8 and 1.6, they will both get floored to 1.  But if we move closer, they will be at 2.1 and 1.9.  So the 2-guy will be in front.  IF the 1.9 guy was added later, then IT will show as being in front when they are at 1.8 and 1.6.  SOo... I gotta fix that..  I can do it by multiplying Z value by 10?  That should do the trick.