SceneScaler - I think it's right afterall..
So as I changed the far distance, my moos were mooving up and down vertically...  But that makes sense, because the vHeight changes..  vHeight doesn't go up as drastically with a farther back rear wall.  I guess that means I need to create an actually logical view OR combine 2, one that stops at horizon, with a tighter view.