movement - Camera, using key presses..
Haven't run it.. needs more first.  need to give camera something to follow.  Need to run update on game objects.  Need to run earlyUpdate on camera..

Camera can either take a location (that can be update with input), or it can take a gameObject, which it'll just track, as that gameObject moves.

InputHandler..  I need to give gameObjects a way to listen to my InputHandler, so I can just call their callbacks when buttonpresses happen..