Day and DayScene - days work!
wooot.  I don't know if I did it stupidly...  I have a Day class which just knows when a day ends.  I have a DayScene which reads that, and then tells the game when a day has ended.  The DayScene fades the game screen, and will have html etc in it, and a button for new day.  Then DayScene will tell the game it's a new day!