Grass - more grass!  Grass grows!!  But daamn it's messed up when it grows
It looks mega whack when it grows on a new day.  Hmm........  It fixes itself when I move the camera, so that means size scaling isn't working right?  Maybe?  Hmmm..

Also not sure how I feel about the not-connected grass pieces..  And the grass looks weird at times as I move.  Double hmm....